Added Call from backend to frontend (Only rated exercises)
Added functionality to call information from the backend to the frontend. 
Due to inconsistency in the database (some questions being rated some not) the feature is disabled in the frontend where further information is commented in. 
Feature can be activated in the future to reduce the load on the network.