add Bamboo workaround with empty commits again
The new version of Bamboo again has the bug (that was previously fixed), that the build plan is not enabled properly and only the second commit leads to an automatic build.
Therefore we need to clone the repository when the student starts the exercise and push an empty commit during initialization.
We now add a variable ‘empty-commit-necessary’ into the yml file to configure this behavior more easily in case the problem is fixed in a future verison.

Signed-off-by: Stephan Krusche <krusche@in.tum.de>