Add additional endpoint to parse data from the bamboo notification plugin.
This adds an additional endpoint (currently /api/resultsnew) that can be used to receive data from the bamboo notification plugin. It parses the request Body and uses it to create the result & feedback.

Some code has already been moved to other methods, but a quite large amount of code still needs to be deduplicated.