remove transactional from several services related to programming exercises
This should improve performance and thread safety.
Our implementation does not really need transactions in the way we use it, in particular if we retrieve objects using repositories and if we save object using repositories.

This needs to be tested on the test server though, because locally I cannot test submission and result retrieval