Add lock and unlock methods to ProgrammingExerciseService
At the moment, those only forward to the "scheduling" profile node, in
the future maybe this should be executed directly on the node, because
this logic can be separated as it does not really have anything to do
with scheduling. This would also allow to send e.g. the number of failed
lock directly and not with a notification.