Improve unnatural phrasing in "created", "updated", and "deleted" strings.
Quite a few files contain these analogous strings, and they all follow a rather awkward sentence structure with an indefinite article and present participle, e.g. "An exercise is updated with identifier {{ param }}". I suggest dropping the article and moving the verb to the front, like so: "Updated exercise with identifier {{ param }}". This is more natural.