Fix misuse of the word "participation"
In a variety of strings the word "participation" is incorrectly used, most commonly by putting it into its nonexistent plural form. I replaced most usages with the word "participant", as this largely captures the same idea. In other instances I reworded the sentence entirely to try and better translate the idea of a "Teilnahme" in German. I am absolutely open for suggestions on how to translate this better!