Update assessment.json
Line 11: A due date cannot be over, as it is a single point in time. Changed "is over" to "has passed"
Line 17: "Must" better conforms to the register of the message than "have to"
Line 19: Same as line 11
Line 48: Rather than saying the due date has not passed yet, I changed "due date" to "submission period" to add emphasis to the fact that submissions are currently ongoing.
Line 50: Add missing dot in ...