Update complaint.json
Line 8: Changed "your total number of complaints" to "this total" to be more concise. It may have been unclear whether "total number of complaints" was supposed to mean the current number of complaints, or the complaint limit.
Lines 9, 12: "may" is more accurate than "can".
Lines 9, 11, 13: Removed superfluous "in total".
Line 12: Changed subject of sentence to "Your team" for clarity.
Line 14: Changed "request" to "complaint" for clarity.
Line 28: "cannot" -> "may not"
Lines 50, 51: Changed "or" to "and" since the two clauses are practically equivalent. Changed "explanations" into "explanation", more natural this way.
Lines 52, 53: "to be unclear to sb."