Update complaintResponse.json
This is something that has caused me and other tutors some confusion in the past. I believe tutors read "Update assessment" and interpret that to mean "send my response", and click it without actually intending to accept the complaint. By changing this to "Accept complaint", I think two things become more clear: the tutor's response will be sent to the student no matter which button is clicked, and clicking this button will effect a change of grade.