Update course.json
Line 14: Removed unnecessary "the". Sounds better this way.
Lines 24, 25, 28, 29: Changed "amount" to "number" (of complaints). "Amount" is not strictly wrong, but "number" is explicitly quantitative.
Lines 25, 29: Reworded explanation to be more concise.
Line 32: Changed "Deadline for complaints" to "Complaint deadline" to be more concise.
Line 33: The adverb "anymore" does not quite fit the register. Changed "complain anymore" to "submit any more complaints" for clarity and to avoid the adverb "anymore".
Line 41: The word "assessments" is in a higher register than "results", and also contains less implied finality (a "result" is unlikely to be changed, but the goal of a complaint is to do exactly that).
Line 45: Changed "without using a complaint" to "without complaining". "Changed "a correction of the assessment" to "assessment corrections". These changes fall in line with natural language.
Line 48: In accordance with line 32, changed "Deadline for more feedback requests" to "More feedback request deadline".
Line 107: Removed unnatural comma, changed "if" to "whether".