Update exercise.json
Line 49: "how often" is not quite the right phrase here, changed to "how many times". Changed "will be added" to "may be added", as this option represents only a limit.
Line 139: Changed "was created unsuccessfully" to "could not be created".
Lines 163, 164, 167, 178: added "number of".
Lines 164, 168: Changed "must be at max" to "may not exceed" to better fit the register.