Update exerciseAssessmentDashboard.json
Line 15: Changed simple past to present perfect because of "so far".
Line 23: Changed simple past to present perfect because of immediacy of action.
Line 31: Fixed incorrect apostrophe placement for plural possessive.
Lines 47, 48, 54, 55: "at this moment" -> "at the moment".
Lines 48, 55: Pluralized "assessment".
Line 82: Changed "assessments of other tutors" to "other tutors' assessments", more natural. Changed present tense to future tense, as the present tense sounds a bit odd.
Line 84: Reworded. Changed "You can assess 0 submissions" to "You may not assess any submissions" as it is much more natural. Removed odd-sounding "in" from "in this round"; works fine as a temporal adverb without the preposition. Changed "are not yet assessed" to "have yet to be assessed" to better conform to the register, and to communicate that those submissions are not only unassessed, but need to be assessed. "from which" -> "of which". Added "currently" for clarity.
Line 86: Removed unnecessary comma. Changed "can" to "may" because the statement is restrictive. Changed present perfect to simple past because of exact timepoint in the past.