Update fileUploadSubmission.json
Line 22: Changed present tense to future tense, as this is only a confirmation question. The temporal adverb "anymore" does not fit the register; reworded sentence to avoid it. Changed "answer" to "submission".
Line 25: "Incorrect" is not quite the right word here; changed to "Unsupported".
Line 26: "Submission for file" -> "Submission of file"
Line 28: Moved verb to the front to be more natural.