Update submission.json
Being able to start a new assessment is not something the user controls, but rather the system. Therefore, it is illogical to ask the user to do something before being able to start a new assessment. Changed "before being able to start a new one" to "before starting a new one".