Update modelingEditor.json
Lines 8, 10: Remove word "still" as it is unnatural in the context. Changed "can" to "may" as the statement is regulatory.
Lines 9, 17: Reword unnatural sentence.
Line 10: Added comma for sentence flow and changed "the score" to "your score".
Line 18: Changed "was submitted" to "is available" to shift the emphasis. I think "is available" more directly communicates that the new assessment was not only submitted by the tutor, but is actually visible to the student.
Line 36: Changed "circle" to "semicircle" (half circle)
Line 48: Reordered sentence for clarity.