Update language in guidedTourCourseExerciseOverview.json
Line 7: Reworded awkwardly structured sentence.
Lines 9, 10: Changed noun "setup" to verb "set up".
Line 10: Changed present tense to future tense; present tense is awkward here.
Line 16: Changed awkward preposition "of" to "for".
Line 20: Changed inaccurate preposition.
Line 32: Reworded for clarity and to correspond with the German translation.
Line 36: Changed "amount" to "number", as it is explicitly quantitative. Changed "start and end time" to "time frame" to be more concise.
Line 40: Changed "See" to "Here you can see" to sound more like a guided tour should. Changed "what" to "which", as there is a limited amount of exercises. Changed "have to be completed" to "are due" to be more direct.
Line 48: Reworded awkward sentence.
Line 52: Changed tense from simple present to present continuous. Changed noun "setup" to verb "set up".
Lines 56, 66, 76: Changed "understand" to "learn". Understanding is surely the goal, but "learn" is the more apt word here.
Line 56: Changed "cancel the tutorial" to "stop the tutorial" to be more natural. Changed "apply the steps that you have seen in the video" to "apply the steps you saw in the video" to be more concise.
Line 61: Changed verb "understand" to "get acquainted with"; I think this sounds much better.
Line 71: Removed unnecessary word "step".
Line 81: Changed "get" to "come". "get" is in a too informal register to be used here.