Update language in guidedTourCourseOverview.json
Line 19: Removed incorrect article; "insight" is uncountable. Changed "in the" to "into" (insight into).
Line 35: Changed incorrect preposition "signed up to" to "signed up for". Changed article "the" to "this", sounds better.
Line 38: Fixed awkward word order by moving "due" to the end.
Line 39: Changed "what" to "which" since there is a limited amount of exercises.
Line 47: Reword to be more concise.