Update language in guidedTourTutorAssessment.json
Line 12: Adjective "received" is awkward in that position. Removed it.
Line 16: Removed "have" (probably a mistake)
Line 20: "exercises of this course" -> "exercises in this course"
Line 41/42: Removed incorrect "like". It is either "what it looks like" or "how it looks".
Line 50: Changed "submissions that have to be assessed" to "submissions to be assessed". Reordered sentence.
Line 54: Removed "you need to", replaced with a comma. Reworded next sentence to be less awkward.
Line 64: Changed present tense to present perfect tense; more natural. Changed "with the point 1" to "with one point".