Update language in student-dashboard.json
Lines 5, 6, 7: Swapped unnatural preposition.
Line 17: Pluralized "courses". Changed "register" to "sign up" to correspond with the text on the button. Fixed preposition.
Line 61: Changed tense to simple past; the quiz is over and cannot be participated in anymore.
Line 94: Changed "reachable points" to more natural "attainable points".
Line 97: Changed "the" to "this". Again, no idea why this sounds better, but it does.
Lines 103, 104: Changed tense to simple past for same reason as in line 61.
Line 106: Changed preposition "of" to "in". Reworded second half of sentence to be more natural.
Lines 107, 109: Reworded awkward sentence. Is still awkward, but less.
Lines 108, 110: "of all exercises" -> "for all exercises"
Line 108: Changed "are" to "is" despite plural noun. English is funny sometimes.
Line 169: Changed simple past to present perfect. Hopefully the question gets answered in the future!
Line 185: Changed "less answers" to "fewer answers".