Change "modeling" to "modelling"
After a code review suggestion from @argertboja I realized that all throughout Artemis the non-standard spelling "modeling" was being used instead of the standard "modelling". I had previously ignored this because I thought that my preference for the spelling with two Ls was an American idiosyncrasy. As it turns out, it's exactly the opposite case. "modelling" is by far the more common spelling internationally.