Rework participant/participation issue
Reverted some changes from commit 66d4b001 and changed some others.
participation.json:
Lines 6, 7: I suggest phrasing it "a student's participation" rather than "a participation" to avoid quantifying it.
Line 9: I'm not particularly happy with "Created new participation" but I suppose it will do.
Lines 10, 11, 13, 16: Changed these lines roughly back to the originals, although I wrote "<>'s participation" instead of "participation of <>" because it sounds better with a student's name.
Line 41: I have spent so long looking at this that I am no longer sure if this sounds weird or not, but I will just write "that participation" and live with it
programmingExercise.json, quizExercise.json:
Changed "participants" to "participation". I do think it would have been clear that the "participants" would only be deleted in the context of the exercise and not globally, but alas. "Participations" is definitely wrong though. Have to leave it in the singular.
exercise.json: Changed "Participants" to "Participation" to correspond with the name of the page as defined in participation.json