Always allow to view the last result if one is currently missing
This is useful because a successful result cannot be identified
otherwise (it has no feedback). Also, #3984 will allow seeing the
commits and requires the result details modal to be available. In case
of missing results, linking the last one to a commit is especially
important to avoid uncertainty.

It might also be worth to ease this "hasFeedback" restriction in other
places for similar reasons.