increase maximum missed class count
The DistributedInstanceMessageSendService no longer gets instrantiated because
now even the SAML2 test runs neeed the scheduling profile.
For all newly added classes proper unit tests have been added. I do not want
to add new tests for random unrelated classes in this branch as it is big enough
as is.