activate 2nd level cache again to improve the performance
Hibernate 6 does not support EAGER collections with @OrderColumn (seems to be a bug), therefore we have to use LAZY here, which is better anyway.

This will lead to some quiz tests failing for now, but we can fix this by loading the collections on our own when the data is needed.

I will try to fix such cases in a follow-up