set a soft maximum length for the detail text
As discussed with Stephan, at TUM nearly all feedbacks are shorter than
1000 characters. Therefore, limiting the feedback detail text to the
shorter length can decrease the table size of the feedback table further
without introducing significant overhead for the separated storage of
the full feedbacks in the long feedback.

Since the SCA feedback is stored as JSON and needs to be able to use the
full length of the available column space and should not create a long
feedback, a separate setDetailTextTruncated has been added to the feedback
to allow setting a feedback where it is guaranteed that no long feedback
will be created.