Add option for graded presentations on grading key page and save data to grading_scale table
This commit adds a new feature to our grading system that allows instructors to activate graded presentations on the grading key page. Instructors can choose between None and Graded presentations, and if Graded is selected, the grading key page now displays the number of presentations and the combined weight of the presentations. This data is then saved to the grading_scale table in the database.
The translations for german and english texts will be added later