Filter Maven Blackbox tests from TemplateIntegrationTest
Maven Blackbox templates were not tested before, as this Project Type is not supported by Bamboo Bitbucket Jira.
Since the TestGroup for ProgrammingExerciseTemplateIntegrationTest changed, the tests were also executed for Maven Blackbox and failed to succeed. We now skip this project type since its templates were not tested before. Nonetheless we should test this project type's templates in the future.