Add private constructor to IrisConstants.java
The implicitly public default constructor was not covered by any tests. Consequently, the jacoco coverage verification failed on the develop branch since the entire IrisConstants.java class is not covered. Making the constructor private mitigates this issue by allowing JaCoCo to detect that there is nothing to cover in this class.