Fix flaky test, please
While I'm not entirely satisfied with the current solution, I've opted for a temporary fix. It's worth noting that other areas also use Files.walk().

This implementation aims to verify if the solution successfully fixes the tests. I will further refine this solution in subsequent commits if the StructuralTestCaseServiceTest(s) succeed.