fix check if project exists in GitLab setups
The previous solution marked a project as existing if any substring of the
project name matched. E.g., a project `COURSE01PROJECTAB` exists. Now
the user wants to create a new exercise with short identifier `COURSE01PROJECTA`.
This was not possible since the search for this identifier also returned
search results matching the different name since they share a common
substring in the name.