check proper equality for plagiarism matches
This fixes #6579.

The plagiarism comparison contains a set of matches. Without an
implementation for proper equality, the same matches were repeated
multiple times and sent like that to the client.
The client then was working correctly in displaying all the relevant
matches, but of course the same text regions were repeated multiple
times once per (repeated) match.