Fix broken html after merge
To take both incoming changes at the merge conflict,
seems to be not correct. The other columns beside the column
for assessment note, does not exist in develop at the moment.
I do not know there these changes are coming from.

The code for the new columns is unchanged by me,
but there are eslint errors in it. Do not kown why there not
noticed earlier (new eslint rules?).
I will fix them in a separate commit.