Limit notes popup to four lines
Notes with much characters are not displayed properly,
so limit the with of the popup and show only the four first lines.
Ending with three dots, when some text is truncated, so indicate that
there is more text.

Formating is changed automaticly by eslint, because the line gets to
long with the addition "popoverClass".