Fix truncated escape sequences not being filtered in C++ template
When truncating the stdout or stderr from a test, we may cut an escape sequence in half, which means that it will not be handled correctly by strip_ansi_codes. To avoid this issues, as a second filtering step, this removes all characters that are not legal inside XML tags.